introduce AnimatedContainer shared component
AnimatedContainer doesn't have a fixed height but is calculated from the children's height. 
Planning to use it for the `Use Existing Wallet Account` optional box.
