+++ app/components/shared/index.js
@@ -30,3 +30,4 @@ export { default as SendTransaction } from "./SendTransaction/SendTransaction";
 export { default as Collapse } from "./Collapse";
 export { default as TicketAutoBuyerForm } from "./TicketAutoBuyerForm";
 export { default as PurchaseTicketsForm } from "./PurchaseTicketsForm";
+export { default as AnimatedContainer } from "./AnimatedContainer";
